+++ CHANGELOG.md
@@ -10,6 +10,7 @@ covered in the changelog.
 ## [Unreleased]
 ### Added
 - `current` command to show currently loaded profiles
+- `--version` flag to display the version of envManager
 
 ## [1.2.0] - 2022-01-26
 ### Added