+++ CHANGELOG.md
@@ -8,6 +8,9 @@ The entries for 1.0.0 and 1.1.0 are reconstructed from the git history, so there
 covered in the changelog.
 
 ## [Unreleased]
+
+
+## [1.3.0] - 2022-02-01
 ### Added
 - `current` command to show currently loaded profiles
 - `--version` flag to display the version of envManager
@@ -26,7 +29,8 @@ covered in the changelog.
 Initial version
 
 
-[Unreleased]: https://github.com/DBX12/envManager/compare/v1.2.0...HEAD
+[Unreleased]: https://github.com/DBX12/envManager/compare/v1.3.0...HEAD
+[1.3.0]: https://github.com/DBX12/envManager/compare/v1.2.0...v1.3.0
 [1.2.0]: https://github.com/DBX12/envManager/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/DBX12/envManager/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/DBX12/envManager/releases/tag/v1.0.0